feat(data): add optional yfinance provider with caching

+++ python/data_provider/base.py
@@ -16,7 +16,7 @@ class OHLCV:
     high: float
     low: float
     close: float
-    volume: float
+    volume: int
 
 
 class DataProvider(Protocol):
